Separates WHDLoad Booter into it's own module. `amiberry_whdbooter.cpp`
